+++ src/main/java/com/notenoughmail/kubejs_tfc/KubeJSTFC.java
@@ -6,8 +6,8 @@ import com.notenoughmail.kubejs_tfc.config.CommonConfig;
 import com.notenoughmail.kubejs_tfc.util.ClientEventHandlers;
 import com.notenoughmail.kubejs_tfc.util.EventHandlers;
 import com.notenoughmail.kubejs_tfc.util.implementation.NamedRegistryWood;
-import com.notenoughmail.kubejs_tfc.util.internal.RockAdder;
-import com.notenoughmail.kubejs_tfc.util.internal.WoodAdder;
+import dev.architectury.platform.Platform;
+import net.dries007.tfc.util.Helpers;
 import net.dries007.tfc.util.registry.RegistryRock;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraftforge.api.distmarker.Dist;
@@ -22,11 +22,57 @@ import java.util.function.Consumer;
 public class KubeJSTFC {
 
     public static final String MOD_NAME = "KubeJS TFC";
-    public static final Logger LOGGER = LogUtils.getLogger();
+    private static final Logger LOGGER = LogUtils.getLogger();
     public static final String MODID = "kubejs_tfc";
 
-    private static Consumer<RockAdder> rockListeners = r -> {};
-    private static Consumer<WoodAdder> woodListeners = w -> {};
+    private static Consumer<ImmutableMap.Builder<String, RegistryRock>> rockListeners = r -> {};
+    private static Consumer<ImmutableMap.Builder<String, NamedRegistryWood>> woodListeners = w -> {};
+
+    public static void info(String message) {
+        LOGGER.info(message);
+    }
+
+    public static void info(String message, Object arg) {
+        LOGGER.info(message, arg);
+    }
+
+    public static void infoLog(String message) {
+        if (Platform.isDevelopmentEnvironment() || Helpers.getValueOrDefault(CommonConfig.debugMode)) {
+            LOGGER.info(message);
+        }
+    }
+
+    public static void infoLog(String message, Object arg) {
+        if (Platform.isDevelopmentEnvironment() || Helpers.getValueOrDefault(CommonConfig.debugMode)) {
+            LOGGER.info(message, arg);
+        }
+    }
+
+    public static void infoLog(String message, Object... args) {
+        if (Platform.isDevelopmentEnvironment() || Helpers.getValueOrDefault(CommonConfig.debugMode)) {
+            LOGGER.info(message, args);
+        }
+    }
+
+    public static void warningLog(String message) {
+        if (Platform.isDevelopmentEnvironment() || Helpers.getValueOrDefault(CommonConfig.debugMode)) {
+            LOGGER.warn(message);
+        }
+    }
+
+    public static void warningLog(String message, Object... args) {
+        if (Platform.isDevelopmentEnvironment() || Helpers.getValueOrDefault(CommonConfig.debugMode)) {
+            LOGGER.warn(message, args);
+        }
+    }
+
+    public static void error(String message) {
+        LOGGER.error(message);
+    }
+
+    public static void error(String message, Object arg) {
+        LOGGER.error(message, arg);
+    }
 
     public KubeJSTFC() {
         EventHandlers.init();
@@ -42,27 +88,25 @@ public class KubeJSTFC {
     }
 
     // Poor man's event bus because scripts are read before the main event bus is started
-    public static void registerRockListener(Consumer<RockAdder> listener) {
+    public static void registerRockListener(Consumer<ImmutableMap.Builder<String, RegistryRock>> listener) {
         rockListeners = rockListeners.andThen(listener);
     }
 
-    public static void registerWoodListener(Consumer<WoodAdder> listener) {
+    public static void registerWoodListener(Consumer<ImmutableMap.Builder<String, NamedRegistryWood>> listener) {
         woodListeners = woodListeners.andThen(listener);
     }
 
     @ApiStatus.Internal
     public static ImmutableMap<String, RegistryRock> registerRocks() {
         final ImmutableMap.Builder<String, RegistryRock> builder = new ImmutableMap.Builder<>();
-        final RockAdder adder = new RockAdder(builder);
-        rockListeners.accept(adder);
+        rockListeners.accept(builder);
         return builder.build();
     }
 
     @ApiStatus.Internal
     public static ImmutableMap<String, NamedRegistryWood> registerWoods() {
         final ImmutableMap.Builder<String, NamedRegistryWood> builder = new ImmutableMap.Builder<>();
-        final WoodAdder adder = new WoodAdder(builder);
-        woodListeners.accept(adder);
+        woodListeners.accept(builder);
         return builder.build();
     }
 }