+++ changelog.md
@@ -1,3 +1,10 @@
+### 1.2.0 (WIP)
+- Add anvil block builder
+- Groundcover block builders can now use preexisting items as their item similar to TFC's sticks
+- Custom farmland and connected grass blocks now have some of TFC's functional tags by default
+- Added event for registering custom 'interactions', see the wiki page for `TFCEvents.registerInteractions`
+- Fully remove legacy container limiting system/event
+
 ### 1.1.2
 - Fix `TFC.misc.rock` and `TFC.misc.rock` maps being empty if called too soon